Fix to info section
It looks like the info block was causing everything after the info block was started to be rendered in the info block. This is just a quick fix to make it so what was previously supposed to be in an info block is a small note bullet point so the rest of the doc isn't rendered in an info block.